Test against a fake in-processs icinga server
A lot of the testing doesn't really flex what a real icinga2 server
does. We can implement a simple fake server with about 250 lines of
Go, then we get all the benefits of really quick feedback and no
dependency on an icinga2 server (which isn't super easy to set up) for
most tests. More complicated tests are still performed against a real
server listening on loopback (if available).